swarmd: No more {agent,manager,node}
