+++ cmd/swarmd/main.go
@@ -1,9 +1,17 @@
 package main
 
 import (
+	"fmt"
+	"net"
 	"os"
+	"os/signal"
+
+	"golang.org/x/net/context"
 
 	"github.com/Sirupsen/logrus"
+	engineapi "github.com/docker/engine-api/client"
+	"github.com/docker/swarm-v2/agent"
+	"github.com/docker/swarm-v2/agent/exec/container"
 	"github.com/docker/swarm-v2/log"
 	"github.com/docker/swarm-v2/version"
 	"github.com/spf13/cobra"
@@ -31,20 +39,150 @@ var (
 				log.L.Fatal(err)
 			}
 			logrus.SetLevel(level)
+
+			v, err := cmd.Flags().GetBool("version")
+			if err != nil {
+				log.L.Fatal(err)
+			}
+			if v {
+				version.PrintVersion()
+				os.Exit(0)
+			}
+		},
+		RunE: func(cmd *cobra.Command, args []string) error {
+			ctx := context.Background()
+			hostname, err := cmd.Flags().GetString("hostname")
+			if err != nil {
+				return err
+			}
+			addr, err := cmd.Flags().GetString("listen-remote-api")
+			if err != nil {
+				return err
+			}
+			addrHost, _, err := net.SplitHostPort(addr)
+			if err == nil {
+				ip := net.ParseIP(addrHost)
+				if ip != nil && (ip.IsUnspecified() || ip.IsLoopback()) {
+					fmt.Println("Warning: Specifying a valid address with --listen-remote-api may be necessary for other managers to reach this one.")
+				}
+			}
+
+			unix, err := cmd.Flags().GetString("listen-control-api")
+			if err != nil {
+				return err
+			}
+
+			managerAddr, err := cmd.Flags().GetString("join-addr")
+			if err != nil {
+				return err
+			}
+
+			forceNewCluster, err := cmd.Flags().GetBool("force-new-cluster")
+			if err != nil {
+				return err
+			}
+
+			hb, err := cmd.Flags().GetUint32("heartbeat-tick")
+			if err != nil {
+				return err
+			}
+
+			election, err := cmd.Flags().GetUint32("election-tick")
+			if err != nil {
+				return err
+			}
+
+			stateDir, err := cmd.Flags().GetString("state-dir")
+			if err != nil {
+				return err
+			}
+
+			caHash, err := cmd.Flags().GetString("ca-hash")
+			if err != nil {
+				return err
+			}
+
+			secret, err := cmd.Flags().GetString("secret")
+			if err != nil {
+				return err
+			}
+
+			engineAddr, err := cmd.Flags().GetString("engine-addr")
+			if err != nil {
+				return err
+			}
+
+			// todo: temporary to bypass promotion not working yet
+			ismanager, err := cmd.Flags().GetBool("manager")
+			if err != nil {
+				return err
+			}
+
+			// Create a context for our GRPC call
+			ctx, cancel := context.WithCancel(context.Background())
+			defer cancel()
+
+			client, err := engineapi.NewClient(engineAddr, "", nil, nil)
+			if err != nil {
+				return err
+			}
+
+			executor := container.NewExecutor(client)
+
+			n, err := agent.NewNode(&agent.NodeConfig{
+				Hostname:         hostname,
+				ForceNewCluster:  forceNewCluster,
+				ListenControlAPI: unix,
+				ListenRemoteAPI:  addr,
+				JoinAddr:         managerAddr,
+				StateDir:         stateDir,
+				CAHash:           caHash,
+				Secret:           secret,
+				Executor:         executor,
+				HeartbeatTick:    hb,
+				ElectionTick:     election,
+				IsManager:        ismanager,
+			})
+			if err != nil {
+				return err
+			}
+
+			if err := n.Start(ctx); err != nil {
+				return err
+			}
+
+			c := make(chan os.Signal, 1)
+			signal.Notify(c, os.Interrupt)
+			go func() {
+				<-c
+				n.Stop(ctx)
+			}()
+
+			go func() {
+				<-n.Ready(ctx)
+				if ctx.Err() == nil {
+					logrus.Info("node is ready")
+				}
+			}()
+
+			return n.Err(context.Background())
 		},
 	}
 )
 
 func init() {
-	mainCmd.PersistentFlags().StringP("log-level", "l", "info", "Log level (options \"debug\", \"info\", \"warn\", \"error\", \"fatal\", \"panic\")")
-	mainCmd.PersistentFlags().StringP("state-dir", "d", "/var/lib/docker/cluster", "State directory")
-	mainCmd.PersistentFlags().StringP("ca-hash", "c", "", "Specifies the remote CA root certificate hash, necessary to join the cluster securely")
-	mainCmd.PersistentFlags().StringP("secret", "s", "", "Specifies the secret token required to join the cluster")
-
-	mainCmd.AddCommand(
-		agentCmd,
-		managerCmd,
-		nodeCmd,
-		version.Cmd,
-	)
+	mainCmd.Flags().BoolP("version", "v", false, "Display the version and exit")
+	mainCmd.Flags().StringP("log-level", "l", "info", "Log level (options \"debug\", \"info\", \"warn\", \"error\", \"fatal\", \"panic\")")
+	mainCmd.Flags().StringP("state-dir", "d", "/var/lib/docker/cluster", "State directory")
+	mainCmd.Flags().StringP("ca-hash", "c", "", "Specifies the remote CA root certificate hash, necessary to join the cluster securely")
+	mainCmd.Flags().StringP("secret", "s", "", "Specifies the secret token required to join the cluster")
+	mainCmd.Flags().String("engine-addr", "unix:///var/run/docker.sock", "Address of engine instance of agent.")
+	mainCmd.Flags().String("hostname", "", "Override reported agent hostname")
+	mainCmd.Flags().String("listen-remote-api", "0.0.0.0:4242", "Listen address for remote API")
+	mainCmd.Flags().String("listen-control-api", "/var/run/docker/cluster/docker-swarmd.sock", "Listen socket for control API")
+	mainCmd.Flags().String("join-addr", "", "Join cluster with a node at this address")
+	mainCmd.Flags().Bool("force-new-cluster", false, "Force the creation of a new cluster from data directory")
+	mainCmd.Flags().Uint32("heartbeat-tick", 1, "Defines the heartbeat interval (in seconds) for raft member health-check")
+	mainCmd.Flags().Uint32("election-tick", 3, "Defines the amount of ticks (in seconds) needed without a Leader to trigger a new election")
+	mainCmd.Flags().Bool("manager", false, "Request initial CSR in a manager role")
 }
